Move large project keys array to projects trait
