+++ src/Library/Benzina/Pump/ProjectsPump.php
@@ -15,84 +15,7 @@ class ProjectsPump extends AbstractPump implements PumpInterface
 {
     use ArrayPumpTrait;
     use DoctrinePumpTrait;
-
-    private const PROJECT_KEYS = [
-        'id',
-        'name',
-        'subtitle',
-        'lang',
-        'currency',
-        'currency_rate',
-        'status',
-        'translate',
-        'progress',
-        'owner',
-        'node',
-        'amount',
-        'mincost',
-        'maxcost',
-        'days',
-        'num_investors',
-        'popularity',
-        'num_messengers',
-        'num_posts',
-        'created',
-        'updated',
-        'published',
-        'success',
-        'closed',
-        'passed',
-        'contract_name',
-        'contract_nif',
-        'phone',
-        'contract_email',
-        'address',
-        'zipcode',
-        'location',
-        'country',
-        'image',
-        'description',
-        'motivation',
-        'video',
-        'video_usubs',
-        'about',
-        'goal',
-        'related',
-        'spread',
-        'reward',
-        'category',
-        'keywords',
-        'media',
-        'media_usubs',
-        'currently',
-        'project_location',
-        'scope',
-        'resource',
-        'comment',
-        'contract_entity',
-        'contract_birthdate',
-        'entity_office',
-        'entity_name',
-        'entity_cif',
-        'post_address',
-        'secondary_address',
-        'post_zipcode',
-        'post_location',
-        'post_country',
-        'amount_users',
-        'amount_call',
-        'maxproj',
-        'analytics_id',
-        'facebook_pixel',
-        'social_commitment',
-        'social_commitment_description',
-        'execution_plan',
-        'sustainability_model',
-        'execution_plan_url',
-        'sustainability_model_url',
-        'sign_url',
-        'sign_url_action',
-    ];
+    use ProjectsPumpTrait;
 
     public function __construct(
         private UserRepository $userRepository,
